server side rendering is in the middle
