+++ client/src/components/Stars.jsx
@@ -4,7 +4,7 @@ import StarsModel from './StarsModel.jsx';
 export default class Stars extends React.Component {
   constructor(props) {
     super(props);
-    console.log('In stars: ', this.props);
+    // console.log('In stars: ', this.props);
     this.state = {
       isLoaded: true,
       ratings: this.props.ratings
@@ -12,11 +12,11 @@ export default class Stars extends React.Component {
     this.ratingsLoaded = this.ratingsLoaded.bind(this);
   }
 
-  componentDidUpdate(prevProps, prevState, snapshot) {
-    if (this.props.ratings !== prevProps.ratings) {
-      this.ratingsLoaded(this.props.ratings);
-    }
-  }
+  // componentDidUpdate(prevProps, prevState, snapshot) {
+  //   if (this.props.ratings !== prevProps.ratings) {
+  //     this.ratingsLoaded(this.props.ratings);
+  //   }
+  // }
 
   ratingsLoaded(ratings) {
     this.setState({
@@ -50,7 +50,7 @@ export default class Stars extends React.Component {
           </div>
           <div>
             <span className="floatLeft50">Check-in</span>
-            <span className="floatRight50"><StarsModel rating={this.state.isLoaded ? this.state.ratings[0].checkin : 0} dimensions='15px' /></span>
+            <span className="floatRight50"><StarsModel rating={this.state.isLoaded ? this.state.ratings[0].check_in : 0} dimensions='15px' /></span>
           </div>
           <div>
             <span className="floatLeft50">Value</span>
