+++ webpack.config.js
@@ -28,7 +28,7 @@ var SERVER_DIR = path.join(__dirname, './server')
 
 const client = {
   // This is the 'root' of our client-side application + where webpack will start bundling things up
-  entry: `${SRC_DIR}/index.jsx`,
+  entry: `${SRC_DIR}/index.js`,
   // This is where the index.html lives and where the bundle.js will go
   output: {
     filename: 'bundle.js',
@@ -43,7 +43,7 @@ const client = {
         include: SRC_DIR,
         loader: 'babel-loader',
         query: {
-          presets: ['react', 'es2015']
+          presets: ['react', 'es2015', 'env']
         }
       }
     ]
